% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{determine_units}
\alias{determine_units}
\title{From a set of input data, attempt to establish the units}
\usage{
determine_units(data)
}
\description{
From a set of input data, attempt to establish the units
}
\keyword{internal}
