% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_types.R
\name{geom_macroplot}
\alias{geom_macroplot}
\title{FIA macroplot}
\usage{
geom_macroplot(
  mapping = NULL,
  data = NULL,
  stat = StatMacroplot,
  position = "identity",
  n = 360,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
A geom representing the four FIA macroplots
}
