% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_types.R
\name{geom_microplot}
\alias{geom_microplot}
\title{FIA microplot}
\usage{
geom_microplot(
  mapping = NULL,
  data = NULL,
  stat = StatMicroplot,
  position = "identity",
  n = 360,
  expand = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
A geom representing the four FIA microplots
}
