% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_types.R
\name{geom_midplot}
\alias{geom_midplot}
\title{FIA midplot}
\usage{
geom_midplot(
  mapping = NULL,
  data = NULL,
  stat = StatMidplot,
  position = "identity",
  n = 360,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
A geom representing the four FIA midplots, sometimes referred to as
"subplots".
}
