% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circles.R
\name{geom_plot_circles}
\alias{geom_plot_circles}
\title{FIA subplot circles of arbitrary radius}
\usage{
geom_plot_circles(
  mapping = NULL,
  data = NULL,
  stat = ggforce::StatCircle,
  position = "identity",
  n = 360,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
This function creates the four FIA subplot circles of a given radius,
and is typically used in the convenience wrappers \code{geom_macroplot()},
\code{geom_midplot()}, and \code{geom_microplot()}, which set the radius to the
specified size for each subplot type.
}
