% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subp_center.R
\name{geom_subp_center}
\alias{geom_subp_center}
\title{FIA subplot centers}
\usage{
geom_subp_center(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
This geom creates splits a set of input points into the four FIA plot centers
}
