% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subp_line.R
\name{geom_subp_line}
\alias{geom_subp_line}
\title{FIA subplot lines}
\usage{
geom_subp_line(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
This geom creates the three lines that eminate from plot center to each of
the outliyng subplots.
}
