% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_circles}
\alias{split_circles}
\title{Splits circles into the FIA arrangement}
\usage{
split_circles(data)
}
\arguments{
\item{data}{A data frame of circle coordinates from the \code{setup_data()} step.}
}
\value{
A data frame with replicated circle coordinates.
}
\description{
This function ingests circle data from \code{setup_data()}, replicates each circle
four times and positions them relative to plot center.
}
\keyword{internal}
