% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_coords}
\alias{split_coords}
\title{Move an arbitrary set of coordinates into their subplot locations}
\usage{
split_coords(data)
}
\arguments{
\item{data}{A data frame containing x and y coordinates, a group column and a
subgroup column indicating the subplot membership of the coordinate.}
}
\description{
Move an arbitrary set of coordinates into their subplot locations
}
\keyword{internal}
