% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_points}
\alias{split_points}
\title{Splits points into the FIA arrangement}
\usage{
split_points(data)
}
\description{
This function ingests point data from \code{setup_data()}, replicates each point
four times and positions them relative to plot center.
}
\keyword{internal}
