% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_points_to_lines}
\alias{split_points_to_lines}
\title{Splits points into lines connecting to the outer subplots}
\usage{
split_points_to_lines(data)
}
\description{
This function ingests point data from \code{setup_data()} and creates three
lines originating from splot center to each of the outer subplots.
}
\keyword{internal}
