% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\docType{data}
\name{split_pos}
\alias{split_pos}
\title{The relative positions of subplots}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\usage{
split_pos
}
\description{
The relative positions of subplots
}
\keyword{internal}
